Escape "+" (plus sign) for Google Search contextual menu item
Google replaces plus signs in queries with spaces, so a query URL like

    http://www.google.com/search?q=New+York+Mets

becomes a search of

    New York Mets

This becomes an issue when the user tries to seach strings such as math
equations, or strings that actually contains plus signs.

Unfortuately, NSString's –stringByAddingPercentEscapesUsingEncoding: does not
convert plus signs. I added a line to manually escape them to %2b.